+++ tests/test_dataset.py
@@ -13,6 +13,7 @@ import pandas as pd
 from surprise import BaselineOnly
 from surprise import Dataset
 from surprise import Reader
+from surprise.dataset import get_dataset_dir
 
 
 random.seed(1)
@@ -236,3 +237,14 @@ def test_build_anti_testset():
         assert r == trainset.global_mean
     expect = trainset.n_users * trainset.n_items
     assert trainset.n_ratings + len(anti) == expect
+
+
+def test_get_dataset_dir():
+    '''Test the get_dataset_dir() function.'''
+
+    os.environ['SURPRISE_DATA_FOLDER'] = '/tmp/surprise_data'
+    assert get_dataset_dir() == '/tmp/surprise_data'
+
+    # Fall back to default
+    del os.environ['SURPRISE_DATA_FOLDER']
+    assert get_dataset_dir() == os.path.expanduser('~' + '/.surprise_data/')